Remove a useless cast
* pchfile_wx.hpp: No longer disable '-Wuseless-cast' for files that
include wx headers. Using '-isystem' for wx include paths disables this
warning for wx headers themselves, and removing the warning pragma here
lets anomalies in lmi code be diagnosed.

* alert_wx.cpp: Resolved a '-Wuseless-cast' warning. The compiler deems
two types to be the same, but that is not obvious to a human, and the wx
API could someday change, making such a cast necessary. Rewrote the code
to deduce the appropriate type.