Move a pragma to a more appropriate place
Now that '-isystem' is used for wx include paths, '-Wcast-function-type'
needn't be ignored for all files that include wx headers. Therefore,
moved the gcc pragma to the one function that this warning identifies.
That function was already documented as 'inherently unsafe', and it is
appropriate to express such a concern in a pragma.