Consolidate and simplify '-Wno-' options for cgicc
Now that there's a set of warning options for cgicc sources, it makes
no sense to specify similar options elsewhere for specific source files
in that set. Removed '-Wno-sign-conversion', which has perhaps been made
unnecessary by '-isystem' usage.

It might be stricter to specify
  -Wno-conversion -Wfloat-conversion
as is done elsewhere in this makefile, apparently on the theory that
  -Wconversion enables -Wfloat-conversion
and therefore perhaps
  -Wno-conversion implies -Wno-float-conversion (?)
in which case '-Wfloat-conversion' might need to be specifically
enabled. The documentation isn't necessarily clear on this point. But
it seems needless to worry about floating-point conversions in cgicc.