Replace non-recursive 'cp' with 'install'
As in the last commit, '-c' is used to mark possible future '-C' use.

In the first 'GNUmakefile' hunk, 'install' replaces the combination of
'cp' and 'chmod'. The '--archive' flag for 'cp' there was specified for
no reason other than that in implies '--preserve', which makes these
files executable; 'install -m 0775' accomplishes the same thing.