Replace 'cp' with 'install' for test prerequisites
Formerly, the '--preserve' option had been used with 'cp' merely out of
habit, even though no attributes of these files needed preservation.
Now, 'install' is used instead, with mode 0664 (rather than the more
conventional 0644, for the same reason that umask is 0002).

Losing '--update' is an unfortunate though minor pessimization, but
'install -C' does not seem to be clearly documented. The coreutils
manual's "install-invocation" chapter warns that it "could result in
redundant copies or attributes that are not reset to the correct
defaults" unless "--user, --group and --mode" are specified, but
specifying those three extra options in order to use '--compare' safely
seems excessive. For now, '-c' (a lowercase option that's ignored) is
used where '--update' had been used previously and '-C' might now be
appropriate.