Use '-isystem' for wx, but not for {xml,xslt}wrapp sources
Using '-isystem' instead of '-I' to find included files is a last
resort, when no other technique works as well. For wx, writing pragmata
in a header like 'pchfile_wx.hpp' has the unfortunate effect of masking
real defects in lmi code that includes it. Nevertheless, added such
pragmata, as comments, to 'pchfile_wx.hpp'; compiling with them enabled
and with this line:
   $(wx_include_paths) \
moved from 'sys_include_directories' to 'lmi_include_directories' in
'workhorse.make' shows that the list of pragmata is complete. For wx,
'-Wzero-as-null-pointer-constant' must be disabled by pragma because the
'-isystem' technique does not suppress the diagnostics it elicits in lmi
code that uses wx macros.

OTOH, added such pragmata to 'xml_xslt_wrapp.cpp', and moved the line
+  /opt/lmi/third_party/src
from 'sys_include_directories' to 'lmi_include_directories' in
'workhorse.make', so that '*.cxx' files are included by means of '-I'.
This '.cpp' file is not a header, and is not included by any other file,
so this change can mask no defect in any other code.

All pragmata have been tested with gcc-8.3 and gcc-10.0, with no attempt
to conditionalize them for any earlier version of gcc that might not
recognize them.