Allow wxGrid to compute census view column sizes more efficiently
Override the recently introduced CanMeasureColUsingSameAttr() to
indicate that there is no need to re-query each row for its own
attribute when the same one can be used for measuring the widths of all
of them.

This results in significant speedup for the censuses with many rows
(more than 2x for the total time of switching to varying width columns
for ~4000 rows, for example).