Use schroot 'directory' type instead of deprecated 'plain'
https://gitlab.com/codelibre/schroot/-/blob/master/NEWS.md
| The plain chroot type has been removed. [...]
| Please replace these with the directory chroot type.

The 'directory' type uses a "profile". The 'default' profile mounts
/home, which is dangerous. Even the 'minimal' profile copies four nss
databases into the chroot, which is needless. Created an 'lmi_profile'
type to meet lmi's needs. For now, the three files it contains are all
empty, but soon its 'fstab' will be customized.

A 'directory' chroot with this empty profile is not identical to a
'plain' chroot: it always runs the '/etc/schroot/setup.d/' scripts
(whereas 'plain' chroots ran none), and therefore it takes about half
a second longer to enter.

Incremented $CHRTVER because this is a significant change.

Incidentally added a couple of comments to '.conf' files. Notably,
the 'setup.config' key is mentioned because it may be handy, though
it's not used because it is apparently deprecated.