Customize /etc/schroot/lmi_profile/fstab
A 'directory' schroot sets up the given mounts on entry, and removes
them on exit. An 8-4-4-4-12 UUID is appended to the apparent mountpoint,
which looks like this on the host:
  /var/cache/schroots bound on
    /run/schroot/mount/lmi_bullseye_2-${UUID}/var/cache/apt/archives
The bind mounts that the 'lmi_setup*.sh' scripts set up should now be
unnecessary, and they'll be removed soon. For the moment, it appears
that they can coexist with the new bind mounts.