Disallow single cell operations when non-current row is selected
When a row different from the current one (as indicated by the grid
cursor) is selected, it can be unclear whether the various operations
applying to a single cell, such as "Edit cell" or "Run cell", apply to
the selected row or the current one.

Resolve this ambiguity by simply disallowing such operations in this
situation.