Simplify logic
The survival limit is a single choice from four elements. For reports,
it's translated to three booleans, one for every choice except the
default (no limit). To ascertain the choice for a composite, there is
no more reasonable method than to pick the last cell's choice. AND-ing
the booleans was a complicated way of achieving the same result.

How should an (integer) age or duration limit be composited? Arguments
could be made for using the highest, or the lowest, but both are wrong
from some point of view, and both introduce the possibility of anomalies
(age 70, dying at age 60, e.g.; or age 80, dying at duration 50, on a
product that matures at age 100). Picking the last cell's value is
simple and robust.