Fix defect introduced 20050114T1947Z: function call with no effect
Although this defect was introduced 20050114T1947Z into CVS, it comes
from an original file predating the lmi epoch, which contained:

    round_interest_rate(tiered_load_amd);

which calculates and discards the rounded value of its (unaffected)
argument.

This change has no effect on any of lmi's 1500 or so regression tests,
but presumably some ancient product reaches this obscure corner of the
code. Gave force to the original intention, which seems reasonable
enough: a dynamic separate-account load is in the nature of a decrement
to the interest rate.

The actual resulting fund increment, SepAcctLoad, has never been, but
should always have been, and soon will be, rounded.