Begin to tackle an observed unit-test failure
Two unit tests succeeded on a multi-user server, then failed when run
by another user. The reason is that a temporary directory was created
with a static name (the same for all users) and not removed. Thus, the
first user's directory remained, and the second user lacked permission
to create and remove files therein.

The approach taken here is to uniquify the directory name. In the shell,
it would be natural to use a command like mktemp(1). C and C++ offer
mktemp(3), but the freebsd manpage says it's "particularly dangerous
from a security perspective", the openbsd linker warns whenever it's
used, and the linux manpage says "Never use mktemp()". Instead, lmi's
unique_filepath() function is used here.

One call to fs::create_directory() is not uniquified--see its inline
comment. That call did not cause a unit-test failure because it was
paired with an fs::remove() call to remove that directory. The directory
that's uniquified in this commit should also be removed, of course, and
will be in the next commit.