Resolve another unit-test failure; remove temporary directories
* name_value_pairs_test.cpp: Removed temporary directory. This resolved
an observed unit-test failure.

* authenticity_test.cpp: Removed temporary directory. The unit test had
not failed, because no attempt was made to create a file there, and
permissions happened to allow cd'ing thither with filesystem::chdir().
Removed the directory anyway, if only because it's good hygiene.

It would alternatively have been possible to uniquify these temporary
directories. That's arguably beneficial because it would make failures
less likely (and the failed test will continue to fail on the server
until the directory owned by 'nemo' is deleted, e.g.:
  sudo rmdir /srv/chroot/whatever/tmp/name_value_pairs_test
). However, uniquification would require adding 'path_utility.o' to
these tests' link commands (and that would only sidestep the failure,
because the offending directory owned by 'nemo' can only be removed
by 'nemo' or by the superuser (or a sudoer)).