Slightly simplify keystrokes in a GUI test
For experienced spreadsheet users who favor the keyboard, Esc is the
natural way to clear a selection.

The purpose of Ctrl-Home here is to move focus to the top row. It also
happens to move to the leftmost column, which doesn't matter: Ctrl-Up
would be as good for this purpose. (Users from the "123" era might favor
End-Up, which works in gnumeric as well as the dominant msw spreadsheet,
but not in libreoffice calc or wxGrid.)

It is not necessary to highlight the top row; Ctrl-Home or Ctrl-Up makes
it the "active" row.