Round each rider charge separately
Some products clearly specify that the charge for each rider is to be
rounded separately, as common sense would suggest anyway.

Commit 785e6f4acf9f of 20200812T171129Z rounded only the total charge
for all riders. That's still desirable in addition to this new commit:
it can ensure that the deduction for all riders combined is an integral
number of cents, even if the new round_rider_charges() is configured
to perform no rounding.