Round changes to specamt
If specamt is required to be an integral number of dollars at all times,
then what does it mean to increase it by $0.97 or decrease it by $.01?
Formerly, lmi did this:
   1000 + 0.97 = 1000.97 // temporarily not whole dollars
   round(1000.97) // forced to whole dollars: maybe 1000, maybe 1001
Now, the increment is rounded before it is applied:
   round(0.97) // maybe $1, maybe $0
which gives a final result that's just as good, but not necessarily the
same. The motivation is to perform only integer arithmetic on values
that are constrained to be integers, which ultimately makes it easy to
replace floating-point values with an integral currency type.