Expunge a footling defect marker [344]
At first glance, this comment seemed to apply to the line that follows
it, which sets 'OldDBOpt'; but that variable is used only in this
function, so the comment wouldn't matter.

However, 'git log -G' suggests that the comment might apply to the whole
function. In that case, it doesn't matter, because the function begins
with an early exit if the basis is not the primary ("current") one.