Don't touch a ledger object before calling its Init()
Moved several assignments to ledger members from SetInitialValues() to
a place in the function that calls it, so that these assignments occur
after the last call to Init(). None of the assigned values was used
between the old and the new locus of assigment (and it wouldn't make
sense to do so, anyway).