Prepare to abstract specified amount
At least three sets of variables store specified amount:
    death_benefits::specamt_     and death_benefits::supplamt_
    AccountValue::ActualSpecAmt  and AccountValue::TermSpecAmt
    LedgerInvariant::SpecAmt     and LedgerInvariant::TermSpecAmt
Those in the ledger class are problematic because they should be
treated as write-only output, but are often used for temporary storage
of evanescent values. Furthermore, ledger data members are generally of
type 'double', and it is most convenient for them all to share the same
data type, but a currency type is preferable for actual calculations.
Therefore, established accessors for these ledger variables, to make it
easier in forthcoming commits to refactor code that uses them for
temporary storage.