Rename a variable
Soon, a congener will be established for the actuarial quantity "lx".
"lx" is the mot juste: it's not "tpx" because it's multiplied by a
"radix", and "Lx" is something different. However, a name like
"partial_mortality_l" may too easily be misread as ending in "_1";
"partial_mortality_lx" is less easily misread. For concinnity, then,
"q" should be "qx".