Simplify
Removed a useless assertion. Copy assignment works even if the target
vector did not already have the intended length.

Obviated an 'else' clause by initializing 'partial_mortality_lx' to the
value it would have assigned. The 'if' clause is rarely executed, so
the extra work (if any) is rarely wasted.

Hoisted the initialization of 'inforce_lives' out of the conditional,
and used its value in the initialization of 'partial_mortality_lx'.

Removed a now-redundant assignment to the first 'partial_mortality_lx'
element.