Calculate and store partial-mortality tpx along with lx
lx is simply tpx times a constant radix, so it's just silly to calculate
lx in one place and then divide it by the radix to get tpx in another,
especially when that other place is an output routine that shouldn't be
performing such calculations. And division is generally to be avoided
where multiplication can be used instead.

Alternatively, lx could have replaced tpx in the monthly trace, which is
the only place where tpx is currently used. But tpx seems preferable
there, and might someday be useful elsewhere.