Call set_partial_mortality() OAOO
Partial mortality is independent of calculation basis, so it was
wasteful and pointless to calculate it once per basis.

The system test would have failed to run if, instead of being moved, the
call to set_partial_mortality() had merely been conditionalized thus:
    if(mce_run_gen_curr_sep_full == RunBasis_)
(as seemed plausible). In that case, the function wouldn't have been
called early enough for solves on a guaranteed basis. One basis must be
"primary", and that's usually 'mce_run_gen_curr_sep_full', but not
always. Such conditionals in production code should instead compare
'RunBasis_' to that "primary" basis.