Resolve a marked defect [339]
The defect marker was spurious: lapses really were reflected, because
of the way 'ItLapsed' was used in GetPartMortQ(). Removed that use of
'ItLapsed', however, because partial-mortality factors are independent
of calculation basis, whereas 'ItLapsed' is not. This makes the
description of the alleged defect accurate, whereas previously it was
not.

However, that description identified no true defect, because the
probability of survival is independent of policy lapse. One might
anticipate a system-test regression, but Ledger::ZeroInforceAfterLapse()
prevents that, by design. Removed an old comment on that function: it
is not possible to know the latest lapse duration among all bases until
after all have been calculated.