Revert "Install 'perf'"
This reverts commit 34fb78dfb7d611c60d87323b92ae1d1986f292b4.

It is generally impossible to install 'perf' in a chroot. The package
name is specific to the (host) kernel version, but that exact package
is unlikely to be available for the chroot, which would generally use
a different release's packages. Building from source might work, but
the apt-get idea in this commit generally won't.