Rework a failing command
This:
  git clone
  git submodule update --init
resulted in
  fatal: not a git repository (or any of the parent directories): .git
whereas manually running this:
  git clone
  cd subdir_created_by_git_clone
  git submodule update --init
succeeded.

The preceding commit (retained because it's valuable in itself) happened
to work around this problem. The present commit resolves the problem.

This alternative was tried:

    clone_opts="$coefficiency --recurse-submodules"
    git clone      "$clone_opts" git://git.savannah.nongnu.org/lmi.git \
      || [...similarly...]

but it failed, with 'set -vx' output [one long line split to fit here]:

+ git clone --jobs=32 --recurse-submodules \
 git://git.savannah.nongnu.org/lmi.git
error: option `jobs' expects a numerical value

...which seems to suggest fragility in git's command-line parser.