Add xmlwrapp submodule and use its sources and headers from it
Add a submodule for this library, using the same version as was
previously used in install_miscellanea.make -- and which doesn't need to
be used there any more, as it is always present in the source tree now.

Don't change the build system yet, i.e. continue using the "unified
build" by including all xmlwrapp sources in xml_xslt_wrapp.cpp, but
explicitly use the relative paths to them instead of copying them to
/opt/lmi/third_party/src directory, which can now be removed from the
list of include directories, where it was quite incongruous.

Do add the path to xmlwrapp headers to the list of include directories,
as there doesn't seem to be any sense in copying them anywhere neither.