Specify a default data directory for CGI
Hardcoded '/opt/lmi/data' as the data directory, as several unit tests
already do. For code actually deployed in production this wouldn't be
good enough, but for CGI it is.

This commit addresses a real problem. Without it, the data directory for
'make cgi_tests' defaulted to '.', the build directory, into which the
makefile copied '/opt/lmi/data/configurable_settings.xml', whereas, for
'make cgi_tests', '/opt/lmi/data/configurable_settings.xml' itself is
specified as a '--data_path=' command-line option. Thus, these similar
targets used different configuration files, which seemed to work as long
as one was a copy of the other. However, running a pc-linux-gnu native
build of lmi altered the configuration file in '/opt/lmi/data', which
was then copied to every build directory. Copying the original msw
version to '/opt/lmi/data' made the 'cli_tests' target work again for
msw, but not the 'cgi_tests' target.

Running a pc-linux-gnu build shouldn't break the lmi configuration file
for *-w64-mingw32, of course; but that's a separate problem.