Force print directory more sensibly
Cf. commit 090d19a80e6a of 20160615T2310Z. If the print directory is
to be forced, force it to the data directory instead of the current
directory. If a data directory was specified with '--data_path=', then
that's a better choice. Otherwise, the data directory defaults to the
current directory, and the former behavior obtains.