Consistently use the same config.guess in all build scripts
This has the advantage of ensuring that all third party libraries use
the same build type and also corrects a problem that could arise when
building wxPdfDoc on a system where Git was configured to perform
automatic EOL translations (i.e. core.autocrlf==true), as using
configu.guess included in wxPdfDoc repository in a shell script resulted
in an error in this case, due to the presence of extraneous CR
characters in it.