Invoke all *-config scripts together in the right place
The right place to do this is 'workhorse.make'. That's where 'wx-config'
has always been invoked. With 'xmlwrapp-config' in 'configuration.make',
this message appeared at the top of any 'make' run:
  make: xmlwrapp-config: No such file or directory
but with it in 'workhorse.make', that no longer occurs. Explanation:
'configuration.make' in included by 'GNUmakefile', before it sets $PATH,
but *-config scripts are architecture-dependent and $PATH is required to
choose which one to invoke. On the other hand, 'GNUmakefile' does not
include 'workhorse.make', but rather invokes it in a recursive 'make'
command, passing it the appropriate $PATH for the chosen architecture.