Add an 'install_posix.sh' script
For the long term (i.e., not in the gwc/ subdirectory), it would make
more sense to break the existing 'install_msw.sh' script into a series
of scripts (akin to 'lmi_setup*.sh'), separating steps like
 - fetch prerequisites
 - build
 - install
 - test
and factoring out the cygwin- and wsl-dependent parts. This new script
is of some practical use meanwhile, and comparing it to 'install_msw.sh'
makes the dependent parts clear.