Fix defect introduced 20201012T2343Z: 'config.guess' for cygwin
Commit 1f4ed3ccf5431 of 20201012T2343Z changed the presumptive location
of 'config.guess', but neglected cygwin.

It is assumed that 'config.guess' is found at one (or more) of the
hardcoded locations, and that it is executable. It does not seem
worthwhile to make this test much more elaborate.

Alternatively, 'install_cygwin.bat' could have copied cygwin's
  /usr/share/libtool/build-aux/config.guess
(curiously, found at the default location for debian rather than for
redhat) to /usr/share/misc/ (as the defective commit cited above did for
'gwc/rh.sh'). However, that would have involved writing a command in the
mutable msw batch language and depending on cygwin's practice never to
change. It seems better to keep 'install_cygwin.bat' minimalistic.