Improve local names
The best name for a lone local variable in a one-line function is 'z'.
Soon a product_file_test::read_lingo_file() function will be added, and
following the naming scheme here replaced by writing
    lingo l(lingo_filename_);
would introduce the name 'l', which is generally avoided because it is
too easily confused with the digit '1'.