Reformat
Database names, formerly limited to 20 characters, can henceforth have
as many as 26 characters. (Add three for a "DB_" prefix.) Reason: most
entities from '.product' files will be moved to '.database' files, and
'.product' names require the extra width. It would alternatively be
possible to rename the widest ones, but it seems wiser to move entities
without simultaneously renaming them.

View this commit with '--ignore-all-space' to confirm that it changes
nothing of substance.

'db_sort.sed' will need to be modified to allow more than one hundred
names under a topic, but that needn't be done now.