Rework temporary assertion for 'PolicyForm'
Asserted equality of 'PolicyForm' as determined by both methods (with
concomitant changes to all proprietary product files). Replaced the
assertion macro with a more informative alternative. Allowed the value
by the old method to equal "{PolicyForm}" because 'sample2*' products
set the value of drop-in text variables to the variable's name enclosed
in braces; perhaps the new method should reproduce that value.

Ultimately the old method and its underpinnings will be eradicated, and
the new method will be used for almost all of class product_data's data
members. For now, it is sufficient to demonstrate the new method with a
single data member, 'PolicyForm', which can now vary by state (among
other axes) in the simple and natural way shown in the last few lines
of sample::sample() in the current version of 'dbdict.cpp'.