Revise lingo enumeration for sample products
Added a new enumerator with the explicit value zero, corresponding to an
empty string, which is important because zero is the default value for
database entities. Removed initializers for other enumerators to show
the effect of such a change: the '.database' and '.lingo' files hold a
different index, but looking up that index in the lingo map yields the
same string.