Prune and reorder '.policy' editor entities [337]
The '.policy' editor is a good idea for its original purpose--to present
a few short strings for convenient browsing and editing--but not for
lengthy footnotes...which were only infrequently and haphazardly added
anyway. Removed all entities except those that name files, and the
insurance company's domicile (which may seem a poor fit here, but would
be a worse fit for '.database'); the rest are to be moved to '.lingo'
anyway. Incidentally, this resolves a marked defect.