Explain why 'handle_exceptions.hpp' is included
Searching for a word in a comment on an include-directive (e.g.,
with vim's 'g*') ideally finds a reason for including a header.
Such is now the case for 'handle_exceptions.hpp'. This is useful
when it's unclear why or whether a header must be included, and
for pruning old include-directives that are no longer needed.