Validate products
The immediate purpose is to iterate over all products and states, in
order to test whether the old '.policy' and new '.lingo' methods produce
the same strings--although this test is otherwise generally useful. This
commit subsumes the throwaway 'odd/test_all_products' branch mentioned
here:
  https://lists.nongnu.org/archive/html/lmi/2020-11/msg00020.html
and that branch may soon be removed.

Replaced an old 'gprof' command-line option with this new one, because
'perf' is generally preferable to 'gprof'.