Perform a transcendental calculation annually instead of monthly
Of course, this calculation should be done OAOO in class InterestRates,
but until there's time for that, this small change may make a noticeable
difference in runtime performance (once the validating assertion is
expunged). Dynamic monthly interest decrements cannot affect this gross
rate, so its value does not change from month to month.