Expunge an unwanted exception class
The GPT server has numerous exception classes of its own. Expunged the
"product rule violated" exception, which was unique in that it was used
outside the server's main TU (because the conditions for throwing it
could not readily be detected in that TU). It harmfully obstructed
merging BasicValues::GPTServerInit() into BasicValues::Init().

Left the other special exception classes alone because they're self-
contained and harmless, though overwrought. The original intention was
to catch numerous specific exceptions, translate them to a numeric error
code, and augment their what() messages with unique contract numbers,
for an lmi-based client-server tax system; but no client was ever
finished.