Read product data files earlier
The immediate motivation is to make rounding rules available earlier,
so that values can be rounded when they are loaded from the database.
This is safe because the selection of data files depends only on the
product, which is determined by ctor arguments; because these XML files
are always read in the same way (in their entirety, regardless of any
other context); and because failure to read any of them causes the ctor
to throw.

SetRoundingFunctors() is similarly safe: it merely sets the function
objects used for rounding, and it depends only on a '.rounding' file.

The appropriate '.policy' and '.database' files have already been read
upstream from this change. The '.funds', '.lingo', and '.strata' files
may as well be read at the same time as '.rounding', for concinnity.