Allow path unit tests to pass when using any root name part
Previously, the check in test_oddities() only passed when the current
root name was "Z:" (which is its default value when using Wine) because
one of the tests compared the result of calling fs::absolute(), which
used the actual root name to complete a non-absolute MSW path, with the
hard-coded string always using "Z:" as the root name.

Fix this by using the actual root name value if available and only
falling back on "Z:" if it isn't, in which case the exact value used
doesn't matter anyhow.

Also update the comment which still mentioned boost::filesystem.