Split a unit-test TU in twain
Commutation functions are a coherent abstraction that is independent
of 7702. In the past, testing 7702 tables in the commutation-functions
unit test was a harmless expedient that introduced no new dependency
(CSO tables were already used to test commutation functions, and 7702
tables used them and a (formerly) hardcoded 7702 interest rate). Now,
however, 7702 tables depend on a 7702 interest rate that must soon
vary according to external circumstances that are not relevant to a
unit test for commutation functions.