Refactor to avoid a header-only implementation
Soon the 7702 interest rate will depend on external considerations,
embodied in a set of TUs that may change during development. A
header-only implementation would induce ongoing changes to the
physical design, which are better avoided.

* antediluvian_stubs.cpp: All further changes will be postdiluvian.

It may seem curious that class Irc7702A and its unit test don't depend
on iglp(). The reason is that a table-driven 7702A implementation does
not need to know the interest or mortality rates implicit in the tables
it uses, so this seeming anomaly actually represents an appropriate
separation of concerns.