Save a sorted list of regressions
Added "$(test_dir)/regressions.tsv" to the 'system_test' recipe. This
TSV file is readable as either a text file or a spreadsheet. Its
contents are sorted in decreasing order by absolute error (primary)
and relative error (secondary), making it easy to see which files have
the severest regressions.

Of course, the datestamped "analysis" file can be manipulated in other
ways by running altered versions of the command in the recipe, without
without re-running the same regression test. For example, this command
sorts in the opposite order, switching the primary and secondary sort
keys, and suppresses lines reporting identical testcases:

< /analysis-[datestamp] sed \
  -e 's/   Summary: max abs diff: /\t/' \
  -e 's/ max rel err:  /\t/' \
  -e'/0\t0/d' \
  | sort --key=3g --key=2g \
  2>&1 |less