dblize() where necessary
In common cases such as this:

-        InvariantValues().SpecAmt[j] = SA;
+        InvariantValues().SpecAmt[j] = dblize(SA);

where a currency amount is to be posted to the ledger, for now it must
be converted to double. In the future, the ledger might hold currency
amounts as 101 cents rather than $1.01, so that ledger entries can be
exact integral values as is ardently desired, even though they remain
of type double for consistency. It will be easy to find and replace all
such instances on that happy day.