Record speed measurements
The last dozen commits cost an extra one or five percent for i686 and
x86_64 respectively. Cumulative penalty compared to the last production
release is very roughly five and ten percent respectively. Branch
valyuta/005 is faster than this, though (particularly because it caches
all product files), and much of it has already been ported to master.