Rework maximum-loan calculation
Reworked AccountValue::SetMaxLoan(), starting from the last version
released into production, along the lines of valyuta/005 . The crucial
difference is whether a maximum-loan formula in product specifications
is to be interpreted as currency or floating-point arithmetic: changed
from the former (recently implemented) to the latter (as in the past).
Thus, intermediate components were recently treated as subject to
separate rounding at each step, but now (once again) rounding is
performed only at the last step, because that seems to be the better
reading of the specifications.

The behavior, in this respect only, should match SHA1 74077f91ce97 of
20210126T2046Z (and of valyuta/005).