Round increments and decrements better
Increments and decrements to real-world accounts are whole cents only.
That ideal is not achievable with floating-point fractional dollars
(it will be with whole cents). In the meantime, this commit approaches
closer to the ideal. For instance, examining "AcctVal" in a realistic
(proprietary) testcase that /opt/lmi/test/regressions.tsv identifies
as the one having the greatest differences, the values in the first
three years are:

new 8994.02999999999883585
old 8994.03000000000247383
new 18632.3600000000005821
old 18632.3600000000005821
new 29144.3100000000013097
old 29144.3100000000049477

In the first pair, ...988 is off by 12, but ...024 is off by 24, so
the "new" result (this commit) is better than the "old" (previously).
In the second pair (for the second year), the values are the same: both
are the closest floating-point value to 18632.36 . In the third year,
the difference is larger, and it's obvious which is more accurate.

Later, in the fifty-eighth year, values are off by a full cent:

new 11272352.3300000000745
old 11272352.3399999793619

which accumulates to fifteen cents by the eightieth year.

If lmi were going to continue representing values as floating-point
dollars, this refinement would seem too precious to bother with.
However, it's soon going to switch to integral cents instead (which
matters much more in other cases than this), so this commit improves
precision as much as it can in order to minimize regressions when that
switch is made.