Round halfway cases consistently to even
The IEEE Standard for Floating-Point Arithmetic (IEEE 754) gives the
rationale: rounding halfway to even avoids bias. Rounding away from zero
is more expensive, because it's not implemented in silicon. Therefore,
lmi avoids std::round(): harming speed in order to introduce (harmful)
bias is not a favorable tradeoff.

* ihs_basicval.cpp: In the example discussed inline, both std::round()
and std::nearbyint() (with FE_TONEAREST) arrive at the same result:
  4.90001 --> 490
  4.89999 --> 490
so this change does not affect the outcome.