Mark experience-rating code for expunction
Experience rating affects several parts of lmi; for now, though, marked
only calculations. This has the effect of changing "NetCOICharge" in the
ledger to zero, as it always should have been (but wasn't) for cases
that don't use experience rating, because its definition:
-    NetCoiCharge = CoiCharge - round_coi_charge().c(retention_charge);
was "CoiCharge" (untouched) minus a "retention charge" that was
meaningless in the absence of experience rating.