Replace a free function with a better general implementation
Replaced local query_into() with product_database::query_into(),
which is better because:

 - using the overload makes query_into() calls uniform;

 - the overload throws if value is not preserved when the double
   in the database is converted to currency (rounded to nearest cent);

 - the local function rounded with 'round_minutiae', silently discarding
   any excess precision in the database value; now, assertions fire if
   the currency amount (rounded to nearest cent) is more precise than
   the datum-specific rounding (here, 'round_minutiae') allows.

Alternative not chosen: pass the rounding function object as an extra
query_into() argument, and use it there to assert appropriate rounding
precision--but that would make the calls non-uniform.

A profoundly different design would use currency<rounding_parameters>
instead of just currency, throughout lmi. Thus, specified-amount limit
might be of some hypothetical type such as:
  currency<1,r_downward> MinRenlSpecAmt;
if it is to be rounded downward to one decimal. Then the "alternative"
above wouldn't require passing the rounding function object, because it
would be encoded in the (hypothetical) currency type. However, currency
types would no longer be simply additive: what would be the result of
  operator+(currency<1,r_downward>,currency<0,r_to_nearest>)
? The answer would depend on the (currency) type into which the result
is to be assigned, which is why lmi specified rounding on assignment:
  currency X = round_however.c(new_value);