Prefer unique_ptr to shared_ptr where appropriate
This is the important change:

* basic_values.hpp: s/shared_ptr/unique_ptr/ for several members.

...which is a little simpler than it would have been if make_shared had
been used: a minor point that seems not to come up in discussions of the
pros and cons of make_shared.

* group_values.cpp: Provide dtors needed to compile and link. They're
not immediately obvious, but the compiler and linker identify their
requirements. The one-time cost of adding these headers amortizes to
nothing. Recompiling this TU whenever any header changes does entail a
modest extra cost.

These changes are the tail that shall not wag the dog:

* antediluvian_stubs.cpp: Provide dtors needed to compile and link.
* basicvalues.cpp: Likewise.

Nobody cares much about the antediluvian branch. It exists only because
it includes some ideas that should be ported to lmi.